refactor(go): reorganize backend model files
