--- backend/internal/models/project.go
@@ -1,12 +1,0 @@
-package models
-
-import "time"
-
-type project struct {
-	Name            string
-	Uuid            string
-	Deadline        time.Time
-	MediaReferences []Media
-	Members         []User
-	Discussion      Discussion
-}
